requirements.txt
django
djangorestframework
pillow
requests
djangorestframework-api-key
psycopg2
